backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
websockets==12.0
aiosqlite==0.19.0
sqlalchemy==2.0.25
openai==1.13.3
python-dotenv==1.0.0
janus==1.0.0
anyio==4.3.0
